pyproject.toml
[project]
name = "uriel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
]
